+++ source/src/Utility.cs
@@ -3,14 +3,11 @@ using System;
 using System.Reflection;
 using TaleWorlds.CampaignSystem;
 using TaleWorlds.Core;
-using TaleWorlds.Engine;
 using TaleWorlds.InputSystem;
 using TaleWorlds.Library;
 using TaleWorlds.Localization;
 using TaleWorlds.MountAndBlade;
 using TaleWorlds.MountAndBlade.View.Screen;
-using ManagedParameters = TaleWorlds.Core.ManagedParameters;
-using ManagedParametersEnum = TaleWorlds.Core.ManagedParametersEnum;
 using Module = TaleWorlds.MountAndBlade.Module;
 
 namespace RTSCamera